requirements.txt
GitPython
python-dotenv